services/board_service/requirements.txt
fastapi[standard]
grpcio
protobuf
sqlalchemy
aiosqlite